Save customer name, address etc when an invoice is created

Data is stored in invoice_recipient table

Approved-by: Janne Matihaldi <janne.matihaldi@vincit.fi>

+++ backend/model-service/src/main/java/fi/hel/allu/model/service/InvoiceService.java
@@ -3,33 +3,51 @@ package fi.hel.allu.model.service;
 import java.time.ZonedDateTime;
 import java.util.List;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import fi.hel.allu.common.exception.NoSuchEntityException;
 import fi.hel.allu.model.dao.ApplicationDao;
 import fi.hel.allu.model.dao.ChargeBasisDao;
+import fi.hel.allu.model.dao.CustomerDao;
+import fi.hel.allu.model.dao.InvoiceRecipientDao;
 import fi.hel.allu.model.dao.InvoiceDao;
+import fi.hel.allu.model.domain.Application;
 import fi.hel.allu.model.domain.ChargeBasisEntry;
+import fi.hel.allu.model.domain.Customer;
 import fi.hel.allu.model.domain.Invoice;
+import fi.hel.allu.model.domain.InvoiceRecipient;
 import fi.hel.allu.model.domain.InvoiceRow;
+
+import java.util.Optional;
 
 /**
  * The service class responsible for managing invoices
  */
 @Service
 public class InvoiceService {
-  private ChargeBasisDao chargeBasisDao;
-  private InvoiceDao invoiceDao;
-  private PricingService pricingService;
-  private ApplicationDao applicationDao;
+  private static final Logger logger = LoggerFactory.getLogger(InvoiceService.class);
+
+  private final ChargeBasisDao chargeBasisDao;
+  private final InvoiceDao invoiceDao;
+  private final PricingService pricingService;
+  private final ApplicationDao applicationDao;
+  private final InvoiceRecipientDao invoiceRecipientDao;
+  private final CustomerDao customerDao;
 
   @Autowired
-  public InvoiceService(ChargeBasisDao chargeBasisDao, InvoiceDao invoiceDao, PricingService pricingService, ApplicationDao applicationDao) {
+  public InvoiceService(ChargeBasisDao chargeBasisDao, InvoiceDao invoiceDao, PricingService pricingService,
+                        ApplicationDao applicationDao, InvoiceRecipientDao invoiceRecipientDao, CustomerDao customerDao) {
     this.chargeBasisDao = chargeBasisDao;
     this.invoiceDao = invoiceDao;
     this.pricingService = pricingService;
     this.applicationDao = applicationDao;
+    this.invoiceRecipientDao = invoiceRecipientDao;
+    this.customerDao = customerDao;
   }
 
   @Transactional(readOnly = true)
@@ -42,7 +60,15 @@ public class InvoiceService {
     List<ChargeBasisEntry> chargeBasisEntries = chargeBasisDao.getChargeBasis(applicationId);
     List<InvoiceRow> invoiceRows = pricingService.toSingleInvoice(chargeBasisEntries);
     ZonedDateTime invoicingDate = applicationDao.getInvoicingDate(applicationId);
-    Invoice invoice = new Invoice(null, applicationId, invoicingDate, false, sapIdPending, invoiceRows);
+    Application application = applicationDao.findById(applicationId);
+    final Optional<Customer> customerOpt = customerDao.findById(application.getInvoiceRecipientId());
+    if (!customerOpt.isPresent()) {
+      throw new NoSuchEntityException("Unable to create invoice as customer is not found");
+    }
+    final Customer customer = customerOpt.get();
+    final InvoiceRecipient invoiceRecipient = new InvoiceRecipient(customer.getType(), customer.getName());
+    final int invoiceRecipientId = invoiceRecipientDao.insert(invoiceRecipient);
+    Invoice invoice = new Invoice(null, applicationId, invoicingDate, false, sapIdPending, invoiceRows, invoiceRecipientId);
     invoiceDao.deleteByApplication(applicationId);
     invoiceDao.insert(applicationId, invoice);
   }
@@ -54,7 +80,16 @@ public class InvoiceService {
    */
   @Transactional(readOnly = true)
   public List<Invoice> findPending() {
-    return invoiceDao.findPending();
+    List<Invoice> invoices = invoiceDao.findPending();
+    invoices.stream().forEach(i -> {
+      Optional<InvoiceRecipient> opt = invoiceRecipientDao.findById(i.getRecipientId());
+      if (opt.isPresent()) {
+        i.setInvoiceRecipient(opt.get());
+      } else {
+        logger.error("Invoice recipient not found with ID {}", i.getRecipientId());
+      }
+    });
+    return invoices;
   }
 
   /**
